refactor(cli): extract shared agent-factory to eliminate code duplication
Move createPlatformAgent, destroyAgent, saveScreenshot, captureScreenshot
into agent-factory.ts. Both do.ts and connect.ts import from shared module.
